+++ iexec-task-api/src/main/java/com/iexec/core/metric/PlatformMetric.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2025 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,10 +30,10 @@ import java.util.LinkedHashMap;
 @JsonDeserialize(builder = PlatformMetric.PlatformMetricBuilder.class)
 public class PlatformMetric {
     int aliveWorkers;
-    int aliveTotalCpu;
-    int aliveAvailableCpu;
-    int aliveTotalGpu;
-    int aliveAvailableGpu;
+    int aliveComputingCpu;
+    int aliveRegisteredCpu;
+    int aliveComputingGpu;
+    int aliveRegisteredGpu;
     LinkedHashMap<TaskStatus, Long> currentTaskStatusesCount;
     long dealEventsCount;
     long dealsCount;